+++ docs/changelog.md
@@ -1,9 +1,15 @@
 # TBD
+### Features
+* Added `nearup` to start a local NEAR network inside the `nearup` Docker image
+
 ### Fixes
 * Pinned base Node image in Dockerfile to `node:14.17.6-alpine3.14`
 
 ### Changes
 * Upgrade to `lambda-api-lib` 0.9.0
+
+### Breaking Changes
+* Add a URL to the result for accessing the `nearcore` daemons started inside the `nearup` Docker container
 
 # 0.2.1
 ### Features